# rest/swaggerui/swagger.yaml
swagger: "2.0"
info:
  version: 0.1.0
  title: Openstack Big Data tools
  description: "Данная система предоставляет REST API для создания кластеров с заданным списком зарегестрированных сервисов в облачной среде OpenStack."
  license:
    name: License Here
  #    url:
  contact:
    email: modis@ispras.ru
host: host.here.com
basePath: /home
schemes:
  - http
  - https
...
paths:
  /projects:
    get:
      tags:
        - projects
      summary: Возвращает список зарегестрированных в системе проектов
      operationId: listProjects
      parameters:
        - name: view
          in: query
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Projects'
    post:
      tags:
        - projects
      summary: Создание нового проекта системы
      description: Создание нового проекта системы
      operationId: createProject
      parameters:
        - name: project
          in: body
          schema:
            $ref: '#/definitions/Project'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Project'
  /projects/{projectId}:
    get:
      tags:
        - projects
      summary: Получение информации о выбранном проекте
      operationId: ProjectGet
      parameters:
        - name: projectId
          description: "ID проекта."
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Project'
    put:
      tags:
        - projects
      summary: Обновление информации о проекте
      description: Обновление информации о проекте
      operationId: updateProject
      parameters:
        - name: project
          in: body
          schema:
            $ref: '#/definitions/Project'
        - name: projectId
          description: "ID проекта."
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Project'
    delete:
      tags:
        - projects
      summary: Удаление проекта системы
      operationId: deleteProject
      parameters:
        - name: projectId
          description: "ID проекта."
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Project'
...
definitions:
...
  Projects:
    type: object
    example:
      {
        "items":
          [
          {
            "ID":"uuid",
            "Name":"projectName",
            "DisplayName":"projectName",
            "GroupId": "id",
            "Description": "someDescription"
          }
          ]
      }
  Project:
    type: object
    example:
      {
        "ID":"uuid",
        "Name":"projectName",
        "DisplayName":"projectName",
        "GroupId": "id",
        "Description": "someDescription"
      }
...